delete contributions.js
