+++ support-frontend/assets/pages/new-contributions-landing/components/ContributionFormContainer.jsx
@@ -9,7 +9,6 @@ import { type ReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
 import React from 'react';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router';
-import { countryGroupSpecificDetails } from 'helpers/internationalisation/contributions.jsx';
 import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
 import { type ErrorReason } from 'helpers/errorReasons';
 import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/readerRevenueApis';
@@ -17,6 +16,7 @@ import { type CreatePaypalPaymentData } from 'helpers/paymentIntegrations/oneOff
 import type { IsoCurrency } from 'helpers/internationalisation/currency';
 import { type LandingPageCopyTestVariant } from 'helpers/abTests/abtestDefinitions';
 import DirectDebitPopUpForm from 'components/directDebit/directDebitPopUpForm/directDebitPopUpForm';
+import { classNameWithModifiers } from 'helpers/utilities';
 import { openDirectDebitPopUp } from 'components/directDebit/directDebitActions';
 
 import { type State } from '../contributionsLandingReducer';
@@ -30,7 +30,6 @@ import {
   setCheckoutFormHasBeenSubmitted,
   createOneOffPayPalPayment,
 } from '../contributionsLandingActions';
-import { classNameWithModifiers } from 'helpers/utilities';
 
 
 // ----- Types ----- //
@@ -82,6 +81,80 @@ const mapDispatchToProps = (dispatch: Function) => ({
   createOneOffPayPalPayment: (data: CreatePaypalPaymentData) => { dispatch(createOneOffPayPalPayment(data)); },
   payPalSetHasLoaded: () => { dispatch(setPayPalHasLoaded()); },
 });
+
+// ----- Functions ----- //
+
+const defaultHeaderCopy = 'Help\xa0us\xa0deliver\nthe\xa0independent\njournalism\xa0the\nworld\xa0needs';
+const defaultContributeCopy = (
+  <span>
+    Make a recurring commitment to support The&nbsp;Guardian long-term or a single contribution
+    as and when you feel like it – choose the option that suits you best.
+  </span>);
+
+const defaultHeaderCopyAndContributeCopy = {
+  headerCopy: defaultHeaderCopy,
+  contributeCopy: defaultContributeCopy,
+};
+
+const helpVariantHeaderCopyAndContributeCopy = {
+  headerCopy: defaultHeaderCopy,
+  contributeCopy: (
+    <span>
+      The Guardian is editorially independent, meaning we set our own agenda. Our journalism is free from commercial
+      bias and not influenced by billionaire owners, politicians or shareholders. No one edits our editor. No one
+      steers our opinion. This is important as it enables us to give a voice to those less heard, challenge the
+      powerful and hold them to account. It’s what makes us different to so many others in the media, at a time when
+      factual, honest reporting is crucial.
+      <span className="bold">Your support is critical for the future of Guardian journalism.</span>
+    </span>
+  ),
+  headerClasses: 'help-variant',
+};
+
+const defaultCountryGroupSpecificDetails = {
+  GBPCountries: defaultHeaderCopyAndContributeCopy,
+  EURCountries: defaultHeaderCopyAndContributeCopy,
+  UnitedStates: defaultHeaderCopyAndContributeCopy,
+  AUDCountries: {
+    ...defaultHeaderCopyAndContributeCopy,
+    headerCopy: 'Help us deliver the independent journalism Australia needs',
+  },
+  International: defaultHeaderCopyAndContributeCopy,
+  NZDCountries: defaultHeaderCopyAndContributeCopy,
+  Canada: defaultHeaderCopyAndContributeCopy,
+};
+
+const helpVariantCountryGroupSpecificDetails = {
+  GBPCountries: helpVariantHeaderCopyAndContributeCopy,
+  EURCountries: helpVariantHeaderCopyAndContributeCopy,
+  UnitedStates: helpVariantHeaderCopyAndContributeCopy,
+  AUDCountries: helpVariantHeaderCopyAndContributeCopy,
+  International: helpVariantHeaderCopyAndContributeCopy,
+  NZDCountries: helpVariantHeaderCopyAndContributeCopy,
+  Canada: helpVariantHeaderCopyAndContributeCopy,
+};
+
+export type CountryMetaData = {
+  headerCopy: string,
+  contributeCopy?: React$Element<string>,
+  headerClasses?: string,
+  // URL to fetch ticker data from. null/undefined implies no ticker
+  tickerJsonUrl?: string,
+};
+
+const countryGroupSpecificDetails: (variant: LandingPageCopyTestVariant) => {
+  [CountryGroupId]: CountryMetaData
+} = (variant: LandingPageCopyTestVariant) => {
+  switch (variant) {
+    case 'control':
+    case 'notintest':
+    default:
+      return defaultCountryGroupSpecificDetails;
+    case 'help':
+      return helpVariantCountryGroupSpecificDetails;
+  }
+};
+
 
 // ----- Render ----- //
